Title: 정수 세 개 입력받아 출력, Time: 117ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240115/정수 세 개 입력받아 출력/take-three-integers-and-output.py
@@ -1,4 +1,4 @@
-a,b=map(float,input().split())
-c=float(input())
+a,b=map(int,input().split())
+c=int(input())
 
 print(a,b,c)
